requirements.txt
ipykernel
nltk
tensorflow
scikit-learn
python-dotenv

# langchain
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
langchain_ollama
langchain_huggingface
sentence_transformers
regex
faiss-cpu # facebook ai sementic search
langchain_chroma
streamlit
langchain_groq

